requirements.txt
comet @ git+https://github.com/chime-experiment/comet.git#egg=comet
kotekan @ git+https://github.com/kotekan/kotekan.git@ael/py3_only#egg=kotekan&subdirectory=python/
aiohttp
pyyaml
requests
sanic>=20.6.0
aioredis>=2.0
async_timeout
redis
prometheus_client<0.8
atomicwrites
mmh3
deepdiff
msgpack
jinja2
